Start fixing code failing the tests in Klaus Dormann's functional test suite
Start fixing code failing the tests in Klaus Dormann's functional test suite

This commit adds some undocumented behavior with the Break and
Expansion flags in the PHP and BRK instructions and a few others.
Use isinstance instead of type
It also cleans up some flake8 and black lints.

For reference:

https://github.com/Klaus2m5/6502_65C02_functional_tests/issues/13
http://forum.6502.org/viewtopic.php?f=2&t=2241&hilit=request+for+verification&start=30#p20914
https://www.nesdev.org/wiki/Visual6502wiki/6502_BRK_and_B_bit