More fixes to get the functional tests passing
Fix off-by-one bugs with pushing and pulling addresses from the stack.
For BRK, push PC + 1, for JSR, push PC - 1, for RTS add one to the
address on the stack address.

Fix some bugs and standardize the N and Z flag updating in the IN* and
DE* instructions.

Move the I flag updating in the BRK instruction.

Remove the I flag clearing in RTI and instead set the E flag and clear
the B flag.