Remove isinstance calls and type-based logic for addressing modes.
There was some usage of isinstance calls to determine where to write
in a few instructions.  The two options were the accumulator and
memory.  This has been refactored to use a write method on
AddressingMode subclasses and the base class.

This required moving IncompleteInstruction out of the instruction.py
file because of circular imports.

Also start a refactor of the other isinstance use for patching bugs.
We'll use a config-based patching system with similar method
calls.